This PR is a rebase of #557 from @SamSaffron

---

The mime types gem (even with the magical columnar store by @jeremyevans) is a memory hog.

In particular on boot, mime/types/columnar will allocate **109k** objects and end up with a retained count of **31K** objects.

This bloat leaves objects that need to be marked and swept every major GC and bloats ruby processes.

To circumvent mini_mime was created.

It uses the exact same database as the full fledged mime types and is capable of only loading stuff on demand with a practical, safe and bound in-memory cache.

mini_mime will allocate **398** objects on boot and only retain **62** (due to rubygems inefficiency)

In table form:

boot allocations | boot retained | lookup | lookup uncached
-- | -- | -- | --
mini_mime | 398 | 62 | 641K/s | 33K/s
mime-types | 109796 | 31165 | 361K/s | 361K/s

The performance of mini_mime is pretty good, cached lookups are faster than the mime types gem and uncached lookups are 10x slower. (which is really not a huge issue considering the memory savings)