@matthewford 2019-01-31
@mattmanning any chance of merging this?